Updated methods according to new type od diagnosedTs field

+++ src/main/java/Model/Tree.java
@@ -54,7 +54,7 @@ public class Tree {
         
     }
     
-    float getRootDiagnosedTs() {
+    long getRootDiagnosedTs() {
     	return root.data.getDiagnosedTs();
     }
